fix: Modified code to deserialize Dispute, Resolve and Chargeback transactions
Those transaction doesn't have field "amount" and to handle that csv reader should
be configured to handle flexible number of fields in row and field "amount" in
InputRecord has to be Option.
Added new error in case that field "amount" is missing for operation that needs that
value.
Modified Engine to handle changed described above.