fix: Modified ReaderBuilder to trim csv input
CSV inputs that had whitespaces and tabs were not parsed correctly.
To fix that ReaderBuilder was configured with Trim::All so that headers
and data records will be trimmed of whitespace characters.
Added UT to check if fix is working.
Added generic builder function to build Reader the same way in main function
and in unit tests.